Fixed a bunch of bugs in our neighbornet implementation.  In some cases it seems to deliver the correct result, although I suspect there could be an issue reducing clusters of 4 vertices into a single cluster with 2 vertices.

+++ core/src/main/java/uk/ac/uea/cmp/spectre/core/ds/split/circular/ordering/CVMatrices.java
@@ -267,9 +267,9 @@ public class CVMatrices {
         return this.c2vs.entrySet();
     }
 
-    public int sumVertex2Components(Identifier v) {
+    public double sumVertex2Components(Identifier v) {
 
-        int sum = 0;
+        double sum = 0.0;
 
         for (Map.Entry<Identifier, Identifier> entry : this.v2c.entrySet()) {
             sum += this.getDistance(entry.getValue(), v);
